requirements.txt
langchain
langchain-groq
langchain-community
langchain_huggingface
python-dotenv
fastapi
faiss-cpu
uvicorn